+++ NEWS.md
@@ -6,9 +6,11 @@
 
 - Make sure `search_index.json` contains valid characters for the JSON format (thanks, @wlandau, #913).
 
-## MINOR CHANGES
+## MAJOR CHANGES
 
 - The `--file-scope` behavior introduced in bookdown v0.20 is now disabled by default. This is due to broken TOC links for duplicate section names (e.g., "Exercises"; see #909) that have automatically generated identifiers.
+
+- The `clean_envir` argument of `bookdown::render_book()` has been deprecated and will be removed in the future (thanks, @jenslaufer, #932).
 
 # CHANGES IN bookdown VERSION 0.20